move tests to another process
