+++ src/js/node/tty.js
@@ -6,10 +6,6 @@ function ReadStream(fd) {
 
   const stream = require("node:fs").ReadStream.call(this, `/dev/fd/${fd}`);
 
-  if (!isatty(stream.fd)) {
-    throw new TypeError("TTY initialization failed: file descriptor is not a TTY");
-  }
-
   stream.isRaw = false;
   stream.isTTY = true;
 
@@ -98,10 +94,6 @@ function WriteStream(fd) {
   if (fd >> 0 !== fd || fd < 0) throw new RangeError("fd must be a positive integer");
 
   const stream = require("node:fs").WriteStream.call(this, `/dev/fd/${fd}`);
-
-  if (!isatty(stream.fd)) {
-    throw new TypeError("TTY initialization failed: file descriptor is not a TTY");
-  }
 
   stream.columns = undefined;
   stream.rows = undefined;
